app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
icalendar
requests
apscheduler
llama-cpp-python
